--- src/main/java/com/example/sns_feed/board/dto/response/BoardSaveResponseDto.java
@@ -1,27 +1,0 @@
-package com.example.sns_feed.board.dto.response;
-
-import com.example.sns_feed.board.entity.Board;
-import jakarta.persistence.GeneratedValue;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Getter
-@RequiredArgsConstructor
-public class BoardSaveResponseDto {
-
-    private final Long id;
-
-    private final String userName;
-
-    private final String title;
-
-    private final String contents;
-
-    private final LocalDateTime createdAt;
-
-    private final LocalDateTime updatedAt;
-
-
-}
