--- src/main/java/com/example/sns_feed/comment/controller/CommentController.java
@@ -1,14 +1,0 @@
-package com.example.sns_feed.comment.controller;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/comments")
-public class CommentController {
-
-    //@PostMapping
-    //public ResponseEntity
-}
