--- src/main/java/com/example/sns_feed/comment/dto/CommentRequestDto.java
@@ -1,13 +1,0 @@
-package com.example.sns_feed.comment.dto;
-
-import lombok.Getter;
-
-@Getter
-public class CommentRequestDto {
-
-    private final String contents;
-
-    public CommentRequestDto(String contents) {
-        this.contents = contents;
-    }
-}
