--- src/main/java/com/example/sns_feed/comment/service/CommentService.java
@@ -1,7 +1,0 @@
-package com.example.sns_feed.comment.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class CommentService {
-}
