+++ src/main/java/com/example/sns_feed/domain/board/SnsFeed.sql
