+++ src/main/java/com/example/sns_feed/domain/board/dto/response/BoardSaveResponseDto.java
@@ -1,0 +1,25 @@
+package com.example.sns_feed.domain.board.dto.response;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@Getter
+@RequiredArgsConstructor
+public class BoardSaveResponseDto {
+
+    private final Long id;
+
+    private final String userName;
+
+    private final String title;
+
+    private final String contents;
+
+    private final LocalDateTime createdAt;
+
+    private final LocalDateTime updatedAt;
+
+
+}
