+++ src/main/java/com/example/sns_feed/domain/comment/service/CommentService.java
@@ -1,0 +1,7 @@
+package com.example.sns_feed.domain.comment.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class CommentService {
+}
