+++ src/main/java/com/example/sns_feed/domain/follow/controller/FollowController.java
@@ -1,0 +1,15 @@
+package com.example.sns_feed.domain.follow.controller;
+
+import com.example.sns_feed.domain.follow.service.FollowService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/follows")
+@RequiredArgsConstructor
+public class FollowController {
+
+    private final FollowService followService;
+
+}
