+++ src/main/java/com/example/sns_feed/domain/follow/dto/FollowListDto.java
@@ -1,0 +1,11 @@
+package com.example.sns_feed.domain.follow.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class FollowListDto {
+    private Long userId;
+    private String email;
+}
