+++ src/main/java/com/example/sns_feed/domain/follow/dto/FollowRequestDto.java
@@ -1,0 +1,8 @@
+package com.example.sns_feed.domain.follow.dto;
+
+import lombok.Getter;
+
+@Getter
+public class FollowRequestDto {
+    private Long followingId;
+}
