+++ src/main/java/com/example/sns_feed/domain/follow/enums/FollowStatus.java
@@ -1,0 +1,6 @@
+package com.example.sns_feed.domain.follow.enums;
+
+public enum FollowStatus {
+    PENDING,
+    ACCEPTED
+}
