+++ src/main/java/com/example/sns_feed/domain/follow/service/FollowService.java
@@ -1,0 +1,14 @@
+package com.example.sns_feed.domain.follow.service;
+
+import com.example.sns_feed.domain.follow.repository.FollowRepository;
+import com.example.sns_feed.domain.user.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class FollowService {
+    private final FollowRepository followRepository;
+    private final UserRepository userRepository;
+
+}
