+++ src/main/java/com/example/sns_feed/domain/user/dto/responsedto/UserResponseDto.java
@@ -1,0 +1,12 @@
+package com.example.sns_feed.domain.user.dto.responsedto;
+
+import lombok.Getter;
+
+@Getter
+public class UserResponseDto {
+    private final Long id;
+
+    public UserResponseDto(Long id) {
+        this.id = id;
+    }
+}
