--- src/main/java/com/example/sns_feed/follow/controller/FollowController.java
@@ -1,15 +1,0 @@
-package com.example.sns_feed.follow.controller;
-
-import com.example.sns_feed.follow.service.FollowService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/follows")
-@RequiredArgsConstructor
-public class FollowController {
-
-    private final FollowService followService;
-
-}
