--- src/main/java/com/example/sns_feed/follow/dto/FollowRequestDto.java
@@ -1,8 +1,0 @@
-package com.example.sns_feed.follow.dto;
-
-import lombok.Getter;
-
-@Getter
-public class FollowRequestDto {
-    private Long followingId;
-}
