--- src/main/java/com/example/sns_feed/follow/enums/FollowStatus.java
@@ -1,6 +1,0 @@
-package com.example.sns_feed.follow.enums;
-
-public enum FollowStatus {
-    PENDING,
-    ACCEPTED
-}
